% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\title{The Table function, a foundation for data science}
\usage{
Table()
}
\description{
This function creates a table "object," so to speak. Initializes an empty `list`.
}
\examples{
Table() \%>\% with_columns()
}
\keyword{Table}
