% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Subsetting by rows}
\usage{
where(tbl, col, fun = "equal to", val)
}
\arguments{
\item{tbl}{the input table}

\item{col}{the name of the column that contains values to subset on}

\item{fun}{a string containing the logical statement}

\item{val}{the value to subset by}
}
\description{
Returns a subset table of the original table, based on some logical
statement.
}
\examples{
Table() \%>\%
   with_columns(list("ColumnA", c(1, 2, 3), "ColumnB", c(3, 2, 1))) \%>\%
   where("ColumnA", "above", 1)

}
\keyword{Table,}
\keyword{columns}
