% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_column.R
\name{with_column}
\alias{with_column}
\title{Adding a column to a table}
\usage{
with_column(tbl, col)
}
\arguments{
\item{tbl}{the input table}

\item{col}{the column}
}
\description{
This function adds columns to the initialized table.
}
\examples{
Table() \%>\% with_column(list("ColumnA", c(1, 1, 1))
}
\keyword{Table,}
\keyword{column}
\keyword{columns,}
