% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_columns.R
\name{with_columns}
\alias{with_columns}
\title{Adding columns to a table}
\usage{
with_columns(tbl, cols)
}
\arguments{
\item{tbl}{the input table}

\item{cols}{the columns}
}
\description{
This function adds columns to the initialized table.
}
\examples{
Table() \%>\% with_columns(list("ColumnA", c(1, 1, 1),
                              "ColumnB", c(2, 2, 2)))
}
\keyword{Table,}
\keyword{columns}
